requirements.txt
streamlit
stegano
pillow